pyproject.toml
[tool.poetry]
name = "clab-api-user"
version = "0.1.0"
description = ""
authors = ["pyhilandjy <hilandjy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.0"
pydantic-settings = "^2.4.0"
sqlalchemy = "^2.0.32"
boto3 = "^1.34.155"
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"
pyjwt = "^2.9.0"
supabase = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
